Agrega ambitos de búsqueda en el modelo de producto.

+++ db/migrate/20230520125927_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[7.0]
   def change
     create_table :products do |t|
       t.string :title
-      t.decimal :price
+      t.float :price
       t.boolean :published
       t.belongs_to :user, null: false, foreign_key: true
